refactor device hook -> context
